remove unused abstraction
